refactor(SynthesisPage): Move components to folder
