+++ packages/client/src/modules/play/playerActions/SynthesisPage/SynthesisContent.styles.tsx
