+++ packages/client/src/modules/play/playerActions/SynthesisPage/SynthesisContent.tsx
@@ -1,14 +1,14 @@
 import Button from "@mui/material/Button";
 import { Box, IconButton } from "@mui/material";
 import { useEffect, useState } from "react";
-import { Typography } from "../../common/components/Typography";
-import { usePlay } from "../context/playContext";
-import { Icon } from "../../common/components/Icon";
-import { Dialog } from "../../common/components/Dialog";
+import { Typography } from "../../../common/components/Typography";
+import { usePlay } from "../../context/playContext";
+import { Icon } from "../../../common/components/Icon";
+import { Dialog } from "../../../common/components/Dialog";
 import { ScenarioNameTextField } from "./SynthesisContent.styles";
-import { useTranslation } from "../../translations/useTranslation";
-import { FlexRow } from "../../common/components/Flex";
-import { useCurrentPlayer } from "../context/hooks/player";
+import { useTranslation } from "../../../translations/useTranslation";
+import { FlexRow } from "../../../common/components/Flex";
+import { useCurrentPlayer } from "../../context/hooks/player";
 
 export { SynthesisScenarioName };
