+++ packages/client/src/modules/play/playerActions/SynthesisPage/index.ts
@@ -1,0 +1,1 @@
+export { SynthesisPage } from "./SynthesisPage";
